chore: tests for ImageListActivity (moved to tools/image_viewer)
